landcover-app/requirements.txt
streamlit
torch
transformers
Pillow
numpy
folium
streamlit-folium 